Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/websockets/Create-Secure-url-with-space.any-expected.txt

CONSOLE MESSAGE: line 49: Invalid url for WebSocket ws://web platform.test:49001/echo

PASS Create Secure WebSocket - Pass a URL with a space - SYNTAX_ERR should be thrown 
